requirements.txt
cudatoolkit=10.2.89=hfd86e86_1
cymem=2.0.5=py38h2531618_0
cython-blis=0.7.4=py38h27cfd23_1
decorator=5.0.7=pypi_0
en-core-web-lg=2.3.1=pypi_0
en-core-web-sm=2.3.1=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
gtts=2.2.2=pypi_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.7.3=py38h06a4308_1
importlib_metadata=3.7.3=hd3eb1b0_1
intel-openmp=2020.2=254
ipython=7.22.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.2=py38_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
murmurhash=1.0.5=py38h2531618_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py38hff7bd54_0
nltk=3.5=pypi_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
parso=0.8.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.2.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
plac=1.1.0=py38_1
playsound=1.2.2=pypi_0
preshed=3.0.2=py38he6710b0_1
prompt-toolkit=3.0.18=pypi_0
ptyprocess=0.7.0=pypi_0
pycparser=2.20=py_2
pygments=2.8.1=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.8=hdb3f193_4
pytorch=1.8.1=py3.8_cuda10.2_cudnn7.6.5_0
readline=8.1=h27cfd23_0
regex=2021.3.17=pypi_0
requests=2.25.1=pyhd3eb1b0_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=py38h06a4308_0
spacy=2.3.5=py38hff7bd54_0
speechrecognition=3.8.1=pypi_0
sqlite=3.35.3=hdfb4753_0
srsly=1.0.5=py38h2531618_0
thinc=7.4.5=py38h9a67853_0
tk=8.6.10=hbc83047_0
torchaudio=0.8.1=py38
torchvision=0.9.1=py38_cu102
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd3eb1b0_0
wasabi=0.8.2=pyhd3eb1b0_0
wcwidth=0.2.5=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
